Models, baselines, restructure, linting
