+++ requirements.lock
@@ -59,6 +59,7 @@ backoff==2.2.1
 bcrypt==4.2.0
     # via paramiko
 beautifulsoup4==4.12.3
+    # via google
     # via nbconvert
 bleach==6.1.0
     # via nbconvert
@@ -76,6 +77,8 @@ cachetools==5.4.0
     # via google-auth
 cads-api-client==1.1.0
     # via cdsapi
+cartopy==0.23.0
+    # via salientsdk
 cdsapi==0.7.0
     # via sheerwater-benchmarking
 certifi==2024.7.4
@@ -131,6 +134,7 @@ dask==2024.7.0
     # via dask-expr
     # via dask-kubernetes
     # via distributed
+    # via salientsdk
     # via sheerwater-benchmarking
 dask-expr==1.1.8
     # via dask
@@ -180,9 +184,12 @@ fsspec==2024.6.1
 gcsfs==2024.6.1
     # via sheerwater-benchmarking
 geopandas==1.0.1
+    # via salientsdk
     # via sheerwater-benchmarking
 gilknocker==0.4.1
     # via coiled
+google==3.0.0
+    # via salientsdk
 google-api-core==2.19.1
     # via google-api-python-client
     # via google-cloud-core
@@ -207,6 +214,7 @@ google-auth-oauthlib==1.2.1
 google-cloud-core==2.4.1
     # via google-cloud-storage
 google-cloud-secret-manager==2.20.1
+    # via salientsdk
     # via sheerwater-benchmarking
 google-cloud-storage==2.18.0
     # via gcsfs
@@ -235,9 +243,11 @@ h11==0.14.0
 h2==4.1.0
     # via httpx
 h5netcdf==1.3.0
+    # via salientsdk
     # via sheerwater-benchmarking
 h5py==3.11.0
     # via h5netcdf
+    # via pvlib
     # via sheerwater-benchmarking
 hpack==4.0.0
     # via h2
@@ -367,6 +377,7 @@ markupsafe==2.1.5
     # via jinja2
     # via nbconvert
 matplotlib==3.9.1
+    # via cartopy
     # via sheerwater-benchmarking
 matplotlib-inline==0.1.7
     # via ipykernel
@@ -396,6 +407,7 @@ nbformat==5.10.4
 nest-asyncio==1.6.0
     # via ipykernel
 netcdf4==1.7.1.post1
+    # via salientsdk
     # via sheerwater-benchmarking
 notebook==7.2.1
     # via sheerwater-benchmarking
@@ -404,8 +416,9 @@ notebook-shim==0.2.4
     # via notebook
 numcodecs==0.13.0
     # via zarr
-numpy==1.26.0
+numpy==1.26.4
     # via bokeh
+    # via cartopy
     # via cftime
     # via contourpy
     # via dask
@@ -417,10 +430,12 @@ numpy==1.26.0
     # via numcodecs
     # via numpy-groupies
     # via pandas
+    # via pvlib
     # via pyarrow
     # via pyogrio
     # via rasterio
     # via rioxarray
+    # via salientsdk
     # via scipy
     # via shapely
     # via sheerwater-benchmarking
@@ -436,6 +451,7 @@ overrides==7.7.0
     # via jupyter-server
 packaging==24.1
     # via bokeh
+    # via cartopy
     # via coiled
     # via dask
     # via distributed
@@ -460,6 +476,8 @@ pandas==2.2.2
     # via dask-expr
     # via flox
     # via geopandas
+    # via pvlib
+    # via salientsdk
     # via xarray
 pandocfilters==1.5.1
     # via nbconvert
@@ -507,6 +525,8 @@ ptyprocess==0.7.0
     # via terminado
 pure-eval==0.2.3
     # via stack-data
+pvlib==0.10.5
+    # via salientsdk
 pyarrow==17.0.0
     # via dask
     # via dask-expr
@@ -535,8 +555,11 @@ pyparsing==3.1.2
     # via pip-requirements-parser
     # via snuggs
 pyproj==3.6.1
+    # via cartopy
     # via geopandas
     # via rioxarray
+pyshp==2.3.1
+    # via cartopy
 pytest==8.3.3
     # via sheerwater-benchmarking
 python-box==7.2.0
@@ -560,6 +583,7 @@ pytz==2024.1
     # via dateparser
     # via multiurl
     # via pandas
+    # via pvlib
 pyyaml==6.0.1
     # via bokeh
     # via dask
@@ -593,8 +617,10 @@ requests==2.32.3
     # via jupyterlab-server
     # via kubernetes
     # via multiurl
+    # via pvlib
     # via pykube-ng
     # via requests-oauthlib
+    # via salientsdk
     # via sheerwater-benchmarking
 requests-oauthlib==2.0.0
     # via google-auth-oauthlib
@@ -621,8 +647,12 @@ s3fs==2024.6.1
     # via sheerwater-benchmarking
 s3transfer==0.10.2
     # via boto3
+salientsdk==0.2.7
+    # via sheerwater-benchmarking
 scipy==1.14.0
     # via flox
+    # via pvlib
+    # via salientsdk
     # via sheerwater-benchmarking
 send2trash==1.8.3
     # via jupyter-server
@@ -631,6 +661,7 @@ setuptools==71.1.0
     # via jupyterlab
     # via rasterio
 shapely==2.0.5
+    # via cartopy
     # via geopandas
 six==1.16.0
     # via asttokens
@@ -659,6 +690,7 @@ tinycss2==1.3.0
     # via nbconvert
 toml==0.10.2
     # via coiled
+    # via salientsdk
 toolz==0.12.1
     # via dask
     # via distributed
@@ -732,6 +764,7 @@ wsproto==1.2.0
     # via httpx-ws
 xarray==2024.6.0
     # via rioxarray
+    # via salientsdk
     # via sheerwater-benchmarking
 xyzservices==2024.6.0
     # via bokeh
