+++ sheerwater_benchmarking/data/__init__.py
@@ -1,4 +1,2 @@
-from .era5 import (era5, single_era5, era5_rolled, era5_agg)
-from .ecmwf import (single_iri_ecmwf, iri_ecmwf, ecmwf_agg, single_iri_ecmwf_dense,
-                    ecmwf_rolled)
+from .era5 import (single_era5, single_era5_cleaned, era5_raw, era5_rolled, era5)
 from .masks import (land_sea_mask)
