+++ sheerwater_benchmarking/data/masks.py
@@ -6,7 +6,7 @@ import xarray as xr
 from sheerwater_benchmarking.utils import cacheable
 
 from sheerwater_benchmarking.utils.secrets import cdsapi_secret
-from sheerwater_benchmarking.utils.data_utils import get_grid
+from sheerwater_benchmarking.utils.general_utils import get_grid
 
 
 @cacheable(data_type='array', cache_args=['grid'])
