+++ sheerwater_benchmarking/models/__init__.py
@@ -1,1 +1,2 @@
 """Forecasting models for the Sheerwater benchmarking project."""
+from .salient import salient_forecast
