+++ sheerwater_benchmarking/utils/__init__.py
@@ -1,4 +1,4 @@
 """Utility functions for benchmarking."""
 from .caching import cacheable
-from .secrets import (cdsapi_secret, ecmwf_secret)
+from .secrets import (cdsapi_secret, ecmwf_secret, salient_auth)
 from .remote import dask_remote
